ADEN-10981 Move addMessageBoxToCollapsedElement into module
